### 2017-11-16 e2de198
Update Tobias's affiliation, and index with order of appearance
