### 2017-11-16 977eca3
Add list of research publications utilized DART
