### 2017-11-16 6c8ccc0
Modifying Memory.hpp
