+++ dart/common/Memory.hpp
@@ -33,14 +33,18 @@
 #define DART_COMMON_MEMORY_HPP_
 
 #include <memory>
+#include <iostream>
 
 namespace dart {
 namespace common {
 
-template <typename _Tp, typename... _Args>
-std::shared_ptr<_Tp> make_aligned_shared(_Args&&... __args);
+template <typename T, typename... Args>
+std::shared_ptr<T> make_aligned_shared(Args&&... args);
 
-template<typename T, typename... Args>
+template <typename T, typename... Args>
+std::shared_ptr<T> make_shared(Args&&... args);
+
+template <typename T, typename... Args>
 std::unique_ptr<T> make_unique(Args&&... args);
 
 } // namespace common
