+++ paper.md
@@ -32,7 +32,7 @@ authors:
    orcid: 0000-0000-0000-0000
    affiliation: 3
  - name: C. Karen Liu
-   orcid: 0000-0000-0000-0000
+   orcid: 0000-0001-5926-0905
    affiliation: 3
 affiliations:
  - name: University of Washington
