### 2017-11-17 8c728a5
Update author order based on Karen's suggestion
