### 2017-11-17 #460 @jslee02
Thank you for initiating this process!

I have a question about ORCID and am asking the question here because I couldn't find an available email address to the editors of JOSS.

Unfortunately, we don't have the ORCID for one of the authors because he is passed away. What could we do in this case? We could make one for him, but I wonder if it would be preferred to just leaving it empty or 0000-0000-0000-0000.
