### 2017-11-17 #460 @arfon
@jslee02 - it's ok to omit the ORCID field for the author who has passed away.
