### 2017-11-17 ef33673
Update Sumit's ORCID and affiliation
also removed Mike's orcid field
