### 2017-11-17 591aa0f
Update description of forward/hybrid dynamics formulations
