+++ paper.bib
@@ -5,6 +5,28 @@
   publisher={Springer}
 }
 
+@article{sohl2001recursive,
+  title={A recursive multibody dynamics and sensitivity algorithm for branched kinematic chains},
+  author={Sohl, Garett A and Bobrow, James E},
+  journal={TRANSACTIONS-AMERICAN SOCIETY OF MECHANICAL ENGINEERS JOURNAL OF DYNAMIC SYSTEMS MEASUREMENT AND CONTROL},
+  volume={123},
+  number={3},
+  pages={391--399},
+  year={2001},
+  publisher={AMERICAN SOCIETY MECHANICAL ENGINEERS}
+}
+
+@article{ploen1999coordinate,
+  title={Coordinate-invariant algorithms for robot dynamics},
+  author={Ploen, Scott R and Park, Frank C},
+  journal={IEEE Transactions on Robotics and Automation},
+  volume={15},
+  number={6},
+  pages={1130--1135},
+  year={1999},
+  publisher={IEEE}
+}
+
 @article{stewart1996implicit,
   title={An implicit time-stepping scheme for rigid body dynamics with inelastic collisions and coulomb friction},
   author={Stewart, David E and Trinkle, Jeffrey C},
@@ -14,4 +36,15 @@
   pages={2673--2691},
   year={1996},
   publisher={Wiley Online Library}
+}
+
+@article{park1995lie,
+  title={A Lie group formulation of robot dynamics},
+  author={Park, Frank C and Bobrow, James E and Ploen, Scott R},
+  journal={The International Journal of Robotics Research},
+  volume={14},
+  number={6},
+  pages={609--618},
+  year={1995},
+  publisher={Sage Publications Sage CA: Thousand Oaks, CA}
 }
