### 2017-11-18 #460 @Kevin-Mattheus-Moerman
I can be editor if @labarba can't take this one
