### 2017-11-18 #460 @Kevin-Mattheus-Moerman
@arfon I'm trying to assign myself as editor. It seems whedon removed lorena but it does not look like it assigned me. Can you see what is going on?
