### 2017-11-29 #460 @jslee02
There was a [correction](https://github.com/dartsim/dart/pull/945) of the affiliations. Let me regenerate the PDF.
