### 2017-12-01 #460 @Kevin-Mattheus-Moerman
@bmagyar I know @danielskatz just pinged you to review over at https://github.com/openjournals/joss-reviews/issues/477,... but if this submission interests you as well it would be great if you could assist with review here too.
