### 2017-12-12 #460 @jslee02
@Kevin-Mattheus-Moerman I'm not trying to involve in the reviewer matching process, but I just wonder if you saw @betatim's [comment](https://github.com/openjournals/joss-reviews/issues/460#issuecomment-345426150) above that he's interested in looking at this submission.
