### 2017-12-12 #500 @whedon
[REVIEW]: DART: Dynamic Animation and Robotics Toolkit
**Submitting author:** @jslee02 (<a href="http://orcid.org/0000-0002-2146-7502">Jeongseok Lee</a>)
**Repository:** <a href="https://github.com/dartsim/dart" target ="_blank">https://github.com/dartsim/dart</a>
**Version:** 6.3.0

---
### REVIEWER 1

---
### REVIEWER 2

---
