### 2017-12-12 #500 @Kevin-Mattheus-Moerman
@betatim I am working to get you added as second reviewer with your own check list. You can start already but I'll let you know when it is all set-up for you too. For the moment only @bmagyar has a checkbox list.
