### 2017-12-16 dartsim/dart#948 @betatim
Help with install instructions
I am trying to install DART in a docker container as part of https://github.com/openjournals/joss-reviews/issues/500

My [`Dockerfile`](https://github.com/betatim/dart/blob/65dff622d99d3a8763ef1df31f2dc1ca3666cbc8/binder/Dockerfile) is pretty standard I think, and I added the commands from the install instructions however it fails with `apt-add-repository command not found`. Do you know what I need to install to get that command?
