### 2017-12-17 dartsim/dart#948 @jslee02
I'm not an expert on Docker, but [it seems `software-properties-common` needs to be installed on the docker image](https://askubuntu.com/a/38035) as
```dockerfile
# line 3
# Set up locales properly
RUN apt-get update && \
    apt-get install --yes --no-install-recommends locales software-properties-common && \
    apt-get purge && \
    apt-get clean && \
    rm -rf /var/lib/apt/lists/*
```
Hope this helps.
