### 2018-01-03 dartsim/dart#949 @bmagyar
[joss-review] Add doi to bibtex entries
Please add the doi (if exists) to the bibtex entries.

https://github.com/openjournals/joss-reviews/issues/500
