### 2018-01-03 #500 @bmagyar
Almost all green from me, excellent submission :+1:
