### 2018-01-03 f4e1a5d
Add DOI to bibtex entries (#950)
