+++ paper/paper.bib
@@ -1,8 +1,9 @@
 @book{featherstone2014rigid,
   title={Rigid body dynamics algorithms},
   author={Featherstone, Roy},
-  year={2014},
-  publisher={Springer}
+  year={2008},
+  publisher={Springer},
+  doi = {10.1007/978-1-4899-7560-7}
 }
 
 @article{sohl2001recursive,
@@ -13,7 +14,8 @@
   number={3},
   pages={391--399},
   year={2001},
-  publisher={AMERICAN SOCIETY MECHANICAL ENGINEERS}
+  publisher={AMERICAN SOCIETY MECHANICAL ENGINEERS},
+  doi = {10.1115/1.1376121}
 }
 
 @article{ploen1999coordinate,
@@ -24,7 +26,8 @@
   number={6},
   pages={1130--1135},
   year={1999},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi = {10.1109/70.817677}
 }
 
 @article{stewart1996implicit,
@@ -35,7 +38,8 @@
   number={15},
   pages={2673--2691},
   year={1996},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi = {10.1002/(SICI)1097-0207(19960815)39:15<2673::AID-NME972>3.0.CO;2-I}
 }
 
 @article{park1995lie,
@@ -46,5 +50,6 @@
   number={6},
   pages={609--618},
   year={1995},
-  publisher={Sage Publications Sage CA: Thousand Oaks, CA}
+  publisher={Sage Publications Sage CA: Thousand Oaks, CA},
+  doi = {10.1177/027836499501400606}
 }
