### 2018-01-04 dartsim/dart#949 @jslee02
Thanks for the suggestion! Done by #950.
