### 2018-01-05 8e30728
Update CMake scripts of examples/tutorials for DART 7
