+++ examples/addDeleteSkels/CMakeLists.txt
@@ -2,7 +2,7 @@ cmake_minimum_required(VERSION 2.8.12)
 
 project(addDeleteSkels)
 
-find_package(DART 6.2.0 REQUIRED COMPONENTS utils-urdf gui CONFIG)
+find_package(DART 7 REQUIRED COMPONENTS io-urdf gui CONFIG)
 
 add_compile_options(-std=c++11)
 
