### 2018-01-05 fbf7f77
Update CHANGELOG.md
