### 2018-01-05 5bbb8ec
Update CHANGELOG.md
