### 2018-01-05 cb6ac01
Update CMake scripts of examples/tutorials for DART 7 (#953)
