### 2018-01-06 092b0f8
Add support of urdf 1.0
