+++ CMakeLists.txt
@@ -24,7 +24,7 @@ project(dart)
 
 set(DART_MAJOR_VERSION "4")
 set(DART_MINOR_VERSION "3")
-set(DART_PATCH_VERSION "6")
+set(DART_PATCH_VERSION "7")
 set(DART_VERSION "${DART_MAJOR_VERSION}.${DART_MINOR_VERSION}.${DART_PATCH_VERSION}")
 set(DART_PKG_DESC "Dynamic Animation and Robotics Toolkit.")
 set(DART_PKG_EXTERNAL_DEPS "flann, ccd, fcl")
@@ -160,7 +160,7 @@ if(NOT BUILD_CORE_ONLY)
   # urdfdom
   find_package(urdfdom QUIET)
   if(urdfdom_FOUND)
-    message(STATUS "Looking for urdfdom - found")
+    message(STATUS "Looking for urdfdom - ${urdfdom_headers_VERSION} found")
   else()
     message(SEND_ERROR "Looking for urdfdom - NOT found, please install liburdfdom-dev")
   endif()
@@ -269,7 +269,9 @@ if(HAVE_BULLET_COLLISION)
 endif()
 
 if(NOT BUILD_CORE_ONLY)
-  set(DART_DEPENDENCIES ${urdfdom_LIBRARIES}
+    message(STATUS "FLANN_LIBRARIES: ${FLANN_LIBRARIES}")
+  set(DART_DEPENDENCIES ${FLANN_LIBRARIES}
+                        ${urdfdom_LIBRARIES}
                         ${TinyXML_LIBRARIES}
                         ${TinyXML2_LIBRARIES}
   )
