+++ dart/config.h.in
@@ -10,6 +10,13 @@
 #define DART_VERSION "@DART_VERSION@"
 #define DART_DESCRIPTION "@DART_PKG_DESC@"
 
+/* urdfdom_headers Version number */
+// We define the version numbers of urdfdom_headers here since it doesn't expose
+// the version numbers itself in source level.
+#define URDFDOM_HEADERS_MAJOR_VERSION @urdfdom_headers_VERSION_MAJOR@
+#define URDFDOM_HEADERS_MINOR_VERSION @urdfdom_headers_VERSION_MINOR@
+#define URDFDOM_HEADERS_PATCH_VERSION @urdfdom_headers_VERSION_PATCH@
+
 #cmakedefine HAVE_NLOPT 1
 #cmakedefine HAVE_IPOPT 1
 #cmakedefine HAVE_SNOPT 1
