### 2018-01-06 1337d55
Update DART 4.3 to be compatible with the latest macOS (#955)
