+++ unittests/CMakeLists.txt
@@ -5,7 +5,7 @@
 # Author(s): Can Erdogan <cerdogan3@gatech.edu>,
 #            Jeongseok Lee <jslee02@gmail.com>
 #
-# Geoorgia Tech Graphics Lab and Humanoid Robotics Lab
+# Georgia Tech Graphics Lab and Humanoid Robotics Lab
 #
 # Directed by Prof. C. Karen Liu and Prof. Mike Stilman
 # <karenliu@cc.gatech.edu> <mstilman@cc.gatech.edu>
@@ -82,4 +82,3 @@ foreach(test ${tests})
   endif(NOT contains)
 
 endforeach(test)
-
