+++ ci/install_osx.sh
@@ -1,24 +1,20 @@
 brew update > /dev/null
 
-PACKAGES='
-git
-cmake
-assimp
-dartsim/dart/fcl
-bullet --with-double-precision
-ode --with-libccd --with-double-precision
-flann
-boost
-eigen
-tinyxml
-tinyxml2
-dartsim/dart/libccd
-nlopt
-dartsim/dart/ipopt
-ros/deps/urdfdom
-ros/deps/urdfdom_headers
-ros/deps/console_bridge
-open-scene-graph
-'
-
-brew install $PACKAGES | grep -v '%$'
+brew install git | grep -v '%$'
+brew install cmake | grep -v '%$'
+brew install assimp | grep -v '%$'
+brew install dartsim/dart/fcl | grep -v '%$'
+brew install bullet --with-double-precision | grep -v '%$'
+brew install ode --with-libccd --with-double-precision | grep -v '%$'
+brew install flann | grep -v '%$'
+brew install boost | grep -v '%$'
+brew install eigen | grep -v '%$'
+brew install tinyxml | grep -v '%$'
+brew install tinyxml2 | grep -v '%$'
+brew install dartsim/dart/libccd | grep -v '%$'
+brew install nlopt | grep -v '%$'
+brew install dartsim/dart/ipopt | grep -v '%$'
+brew install ros/deps/urdfdom | grep -v '%$'
+brew install ros/deps/urdfdom_headers | grep -v '%$'
+brew install ros/deps/console_bridge | grep -v '%$'
+brew install open-scene-graph | grep -v '%$'
