### 2018-01-08 #500 @Kevin-Mattheus-Moerman
@jslee02 I have some minor comments on your paper.

1. Could you rephrase the heading "Research Publications Utilized DART"? Perhaps "Research publications utilizing DART" or "Research publications which utilized DART" sounds better.

2. Check the sentence: "In addition, DART gives provides flexibility...", Looks like removal of the word "gives" would fix this.
