### 2018-01-08 9f1b743
Add JOSS paper
