### 2018-01-11 aac8bfd
Fix various build issues with Visual Studio 2017 (15.5.3)
