### 2018-01-11 6050a9d
Fix various build issues with Visual Studio 2017 (15.5.3) (reverted from commit aac8bfd59081a19ea484d6ee5517ec0f0344a83c)
