+++ dart/math/Geometry.cpp
@@ -1591,8 +1591,8 @@ Eigen::Vector2d computeCentroidOfHull(const SupportPolygon& _convexHull)
 
     if(BEYOND_ENDPOINTS == result)
     {
-      double a1 = atan2( (p1-p0)[1], (p1-p0)[0] )*180.0/constantsd::pi();
-      double a2 = atan2( (p2-p0)[1], (p2-p0)[0] )*180.0/constantsd::pi();
+      double a1 = atan2( (p1-p0)[1], (p1-p0)[0] )*180.0/M_PI;
+      double a2 = atan2( (p2-p0)[1], (p2-p0)[0] )*180.0/M_PI;
       double diff = a1-a2;
       dtwarn << "[computeCentroidOfHull] You have passed in a set of points "
              << "which is not a proper convex hull! The invalid segment "
