+++ dart/planning/Path.cpp
@@ -39,7 +39,6 @@
 #include <algorithm>
 #include <cmath>
 #include <Eigen/Geometry>
-#include "dart/math/Constants.hpp"
 
 using namespace std;
 using namespace Eigen;
@@ -149,13 +148,12 @@ public:
 	}
 
 	list<double> getSwitchingPoints() const {
-        const double pi = math::constantsd::pi();
 		list<double> switchingPoints;
 		const double dim = x.size();
 		for(unsigned int i = 0; i < dim; i++) {
 			double switchingAngle = atan2(y[i], x[i]);
 			if(switchingAngle < 0.0) {
-				switchingAngle += pi;
+				switchingAngle += M_PI;
 			}
 			const double switchingPoint = switchingAngle * radius;
 			if(switchingPoint < length) {
