+++ dart/common/Resource.hpp
@@ -35,7 +35,6 @@
 
 #include <cstddef>
 #include <memory>
-#include <string>
 
 namespace dart {
 namespace common {
