### 2018-01-11 8cfb680
Fix various build issues with Visual Studio 2017 (15.5.3)
