+++ CMakeLists.txt
@@ -3,7 +3,7 @@
 #===============================================================================
 # CMake 2.8.12 or above is required for CMakeParseArguments.
 if(MSVC)
-  cmake_minimum_required(VERSION 3.1.3)
+  cmake_minimum_required(VERSION 3.8.0)
 else()
   cmake_minimum_required(VERSION 2.8.12)
 endif()
@@ -159,17 +159,27 @@ if(MSVC)
 
   # Visual Studio enables c++11 support by default
   if(MSVC_VERSION VERSION_LESS 1900)
-    message(FATAL_ERROR "${PROJECT_NAME} requires VS 2015 or greater.")
+    message(FATAL_ERROR "${PROJECT_NAME} requires VS2017 or greater.")
   endif()
   if(DART_TREAT_WARNINGS_AS_ERRORS)
     set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} /WX")
   endif()
-  set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} /MP4")
-  set(CMAKE_EXE_LINKER_FLAGS_RELEASE "/LTCG /INCREMENTAL:NO")
+  set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} /EHsc /permissive-")
+  #set(CMAKE_EXE_LINKER_FLAGS_RELEASE "/LTCG /INCREMENTAL:NO")
   if(NOT DART_MSVC_DEFAULT_OPTIONS)
     set(CMAKE_CXX_FLAGS_DEBUG "${CMAKE_CXX_FLAGS_DEBUG} ${DART_RUNTIME_LIBRARY}d /Zi /Gy /W1 /EHsc")
     set(CMAKE_CXX_FLAGS_RELEASE "${CMAKE_CXX_FLAGS_RELEASE} ${DART_RUNTIME_LIBRARY} /Zi /GL /Gy /W1 /EHsc")
   endif(NOT DART_MSVC_DEFAULT_OPTIONS)
+  add_definitions(-D_CRT_SECURE_NO_WARNINGS)
+  add_compile_options(/wd4334)
+  add_compile_options(/wd4267)
+  add_compile_options(/wd4244)
+  add_compile_options(/wd4250)
+  add_compile_options(/wd4996)
+  add_compile_options(/wd4099)
+  add_compile_options(/wd4305)
+  add_compile_options(/wd4838)
+  add_compile_options(/bigobj)
 
 elseif(CMAKE_COMPILER_IS_GNUCXX)
 
