+++ .appveyor.yml
@@ -1,0 +1,81 @@
+# Specify version format
+version: "{build}"
+
+image: Visual Studio 2017
+
+platform: x64
+
+# specify custom environment variables
+environment:
+  MSVC_DEFAULT_OPTIONS: ON
+  CMAKE_TOOLCHAIN_FILE: -DCMAKE_TOOLCHAIN_FILE="C:\tools\vcpkg\scripts\buildsystems\vcpkg.cmake"
+  matrix:
+    - CMAKE_GENERATOR: -G"Visual Studio 15 2017 Win64"
+      CMAKE_TOOLCHAIN: -T"v141"
+
+# build configuration, i.e. Debug, Release, etc.
+configuration:
+  - Debug
+  - Release
+
+# scripts that are called at very beginning, before repo cloning
+init:
+  - cmd: cmake --version
+  - cmd: msbuild /version
+
+# clone directory
+clone_folder: C:\projects\dart
+
+# branches to build
+branches:
+  # whitelist
+  # only: 
+  #   - master
+  # blacklist
+  except:
+    - gh-pages
+
+# scripts that run after cloning repository
+install:
+  - if "%platform%"=="Win32" set VCPKG_ARCH=x86-windows
+  - if "%platform%"=="x64"   set VCPKG_ARCH=x64-windows
+
+  # install pacakges
+  # ---- required dependencies: boost, eigen3, libccd, fcl, assimp
+  - vcpkg install boost:%VCPKG_ARCH%
+  - vcpkg install eigen3:%VCPKG_ARCH%
+  - vcpkg install ccd:%VCPKG_ARCH%
+  - vcpkg install fcl:%VCPKG_ARCH%
+  - vcpkg install assimp:%VCPKG_ARCH%
+  # ---- optional dependencies: nlopt, ipopt, bullet3, ode, flann, tinyxml2, urdfdom_headers, urdfdom, opengl, freeglut, osg
+  #- vcpkg install nlopt:%VCPKG_ARCH%
+  #- vcpkg install ipopt:%VCPKG_ARCH%
+  #- vcpkg install bullet3:%VCPKG_ARCH%
+  #- vcpkg install ode:%VCPKG_ARCH%
+  #- vcpkg install flann:%VCPKG_ARCH%
+  #- vcpkg install tinyxml2:%VCPKG_ARCH%
+  #- vcpkg install urdfdom_headers:%VCPKG_ARCH%
+  #- vcpkg install urdfdom:%VCPKG_ARCH%
+  #- vcpkg install opengl:%VCPKG_ARCH%
+  #- vcpkg install freeglut:%VCPKG_ARCH%
+  #- vcpkg install osg:%VCPKG_ARCH%
+  
+  - vcpkg integrate install
+
+  - cmd: cd C:\projects\dart
+  - cmd: md build
+  - cmd: cd build
+  - cmd: cmake %CMAKE_GENERATOR% -DCMAKE_BUILD_TYPE=%configuration% -DDART_MSVC_DEFAULT_OPTIONS="%MSVC_DEFAULT_OPTIONS%" %CMAKE_TOOLCHAIN_FILE% %CMAKE_TOOLCHAIN% ..
+  
+# preserve contents of selected directories and files across project builds
+cache:
+  - C:\tools\vcpkg\installed -> .appveyor.yml
+  - C:\tools\vcpkg\packages -> .appveyor.yml
+
+build:
+  project: C:\projects\dart\build\dart.sln # path to Visual Studio solution or project
+  parallel: true                           # enable MSBuild parallel builds
+  verbosity: quiet                         # MSBuild verbosity level (quiet|minimal|normal|detailed)
+
+#test_script:
+  #- cmd: ctest --build-config %configuration% --parallel 4 --output-on-failure
