### 2018-01-11 26c90d0
Fix various build issues with Visual Studio 2017 (15.5.3) (reverted from commit 8cfb6809b855e747c28778087bea237589bedeef)
