+++ dart/math/Geometry.hpp
@@ -36,7 +36,6 @@
 #include <Eigen/Dense>
 
 #include "dart/common/Deprecated.hpp"
-#include "dart/math/Constants.hpp"
 #include "dart/math/MathTypes.hpp"
 
 namespace dart {
@@ -462,9 +461,7 @@ bool verifyTransform(const Eigen::Isometry3d& _T);
 /// rotations
 inline double wrapToPi(double angle)
 {
-  constexpr auto pi = constantsd::pi();
-
-  return std::fmod(angle+pi, 2*pi) - pi;
+  return std::fmod(angle+M_PI, 2*M_PI) - M_PI;
 }
 
 template <typename MatrixType, typename ReturnType>
