### 2018-01-11 bbd6be3
Fix various build issues with Visual Studio 2017 (15.5.3)
