+++ dart/common/SharedLibrary.hpp
@@ -39,11 +39,25 @@
 #include "dart/common/Platform.hpp"
 
 #if DART_OS_LINUX
+
 #define DYNLIB_HANDLE void*
+
 #elif DART_OS_MACOS
+
 #define DYNLIB_HANDLE void*
+
 #elif DART_OS_WINDOWS
+
+#ifdef NOMINMAX
+#include <windows.h>
+#else
+#define NOMINMAX
+#include <windows.h>
+#undef NOMINMAX
+#endif
+using hInstance = HINSTANCE__*;
 #define DYNLIB_HANDLE hInstance
+
 #endif
 
 namespace dart {
