+++ .appveyor.yml
@@ -54,7 +54,7 @@ install:
   - vcpkg install ode:%VCPKG_ARCH%
   - vcpkg install flann:%VCPKG_ARCH%
   - vcpkg install tinyxml2:%VCPKG_ARCH%
-  - vcpkg install urdfdom_headers:%VCPKG_ARCH%
+  - vcpkg install urdfdom-headers:%VCPKG_ARCH%
   - vcpkg install urdfdom:%VCPKG_ARCH%
   - vcpkg install opengl:%VCPKG_ARCH%
   - vcpkg install freeglut:%VCPKG_ARCH%
