### 2018-01-12 a110a87
Install only required dependencies
