### 2018-01-12 fa30432
Fix various build issues with Visual Studio 2017 (15.5.3) (#956)
