### 2018-01-15 dartsim/dart#957 @costashatz
Add publications to DART's Academic Research page
Hello,

I was checking for something in your docs and found the *Academic Research* page (I wasn't aware of this doc page): http://dartsim.github.io/use_cases.html

There you have quite a few of publications that use DART. In our research lab (http://www.resibots.eu/), we have used DART in the following publications:

**Peer-reviewed International Journals and Conferences**

- Chatzilygeroudis, K., & Mouret, J. B. (2018). [Using Parameterized Black-Box Priors to Scale Up Model-Based Policy Search for Robotics](https://arxiv.org/pdf/1709.06917). *Proceedings of the International Conference on Robotics and Automation (ICRA)*.
- Pautrat, R., Chatzilygeroudis, K., & Mouret, J.-B. (2018). [Bayesian Optimization with Automatic Prior Selection for Data-Efficient Direct Policy Search](https://arxiv.org/pdf/1709.06919). *Proceedings of the International Conference on Robotics and Automation (ICRA)*.
- Supratik P., Chatzilygeroudis K., Ciosek K., Mouret JB., Osborne M., and Whiteson Sh. (2018). [Alternating Optimisation and Quadrature for Robust Control](http://www.cs.ox.ac.uk/people/shimon.whiteson/pubs/paulaaai18.pdf). *Proceedings of AAAI Conference on Artificial Intelligence*.
- Chatzilygeroudis, K., Vassiliades, V. and Mouret, J.-B. (2017). [Reset-free Trial-and-Error Learning for Robot Damage Recovery](https://arxiv.org/pdf/1610.04213). *Robotics and Autonomous Systems*.
- Vassiliades V., Chatzilygeroudis K., and Mouret JB. (2016). [Using centroidal voronoi tessellations to scale up the multi-dimensional archive of phenotypic elites algorithm](https://arxiv.org/pdf/1610.05729.pdf). *IEEE Transactions on Evolutionary Computation*.

**Minimally reviewed ones (workshops)**

- Chatzilygeroudis, K., Cully, A. and Mouret, J.-B. (2016). [Towards semi-episodic learning for robot damage recovery](https://arxiv.org/abs/1610.01407). *Workshop on AI for Long-Term Autonomy at the IEEE International Conference on Robotics and Automation 2016*.
- Papaspyros, V., Chatzilygeroudis, K., Vassiliades, V., and Mouret, J.-B. (2016). [Safety-Aware Robot Damage Recovery Using Constrained Bayesian Optimization and Simulated Priors](https://arxiv.org/pdf/1611.09419v3). *Workshop on Bayesian Optimization at the Annual Conference on Neural Information Processing Systems (NIPS) 2016.*

It'd be nice to include (some of) them in the DART docs.

Thanks in advance!
