### 2018-01-16 dartsim/dart#957 @jslee02
Thanks for introducing publications using DART. Awesome!

I added three of them to the [website](http://dartsim.github.io/use_cases.html) because I couldn't find that DART is cited in the other four papers. :sweat:
