### 2018-01-16 dartsim/dart#957 @costashatz
> I added three of them to the website because I couldn't find that DART is cited in the other four papers.

That would be by mistake.. I will update the publications and let you know (I will also give you some pictures to put :smile: )..
