### 2018-01-17 79c130a
Fix for #734
