### 2018-01-17 652957c
Add vector version of joint position limit functions
