### 2018-01-18 15846aa
Apply lazy evaluation to shape volume computation
