+++ dart/dynamics/LineSegmentShape.hpp
@@ -111,7 +111,7 @@ public:
 protected:
 
   // Documentation inherited
-  void updateVolume() override;
+  void updateVolume() const override;
 
   /// Line thickness for rendering
   float mThickness;
