+++ dart/dynamics/BoxShape.cpp
@@ -36,18 +36,22 @@
 namespace dart {
 namespace dynamics {
 
+//==============================================================================
 BoxShape::BoxShape(const Eigen::Vector3d& _size)
   : Shape(BOX),
-    mSize(_size) {
+    mSize(_size)
+{
   assert(_size[0] > 0.0);
   assert(_size[1] > 0.0);
   assert(_size[2] > 0.0);
   mBoundingBox.setMin(-_size * 0.5);
   mBoundingBox.setMax(_size * 0.5);
-  updateVolume();
 }
 
-BoxShape::~BoxShape() {
+//==============================================================================
+BoxShape::~BoxShape()
+{
+  // Do nothing
 }
 
 //==============================================================================
@@ -82,14 +86,16 @@ Eigen::Matrix3d BoxShape::computeInertia(const Eigen::Vector3d& size,
   return inertia;
 }
 
-void BoxShape::setSize(const Eigen::Vector3d& _size) {
+//==============================================================================
+void BoxShape::setSize(const Eigen::Vector3d& _size)
+{
   assert(_size[0] > 0.0);
   assert(_size[1] > 0.0);
   assert(_size[2] > 0.0);
   mSize = _size;
   mBoundingBox.setMin(-_size * 0.5);
   mBoundingBox.setMax(_size * 0.5);
-  updateVolume();
+  mIsVolumeDirty = true;
 }
 
 const Eigen::Vector3d& BoxShape::getSize() const {
@@ -103,9 +109,10 @@ Eigen::Matrix3d BoxShape::computeInertia(double mass) const
 }
 
 //==============================================================================
-void BoxShape::updateVolume()
+void BoxShape::updateVolume() const
 {
   mVolume = computeVolume(mSize);
+  mIsVolumeDirty = false;
 }
 
 }  // namespace dynamics
