+++ dart/dynamics/SphereShape.hpp
@@ -70,7 +70,7 @@ public:
 
 protected:
   // Documentation inherited.
-  void updateVolume() override;
+  void updateVolume() const override;
 
 private:
   /// Radius of this Sphere
