+++ dart/dynamics/CapsuleShape.hpp
@@ -84,7 +84,7 @@ public:
 protected:
 
   // Documentation inherited.
-  void updateVolume() override;
+  void updateVolume() const override;
 
 private:
 
