### 2018-01-18 097d6fb
Apply lazy evaluation to shape bounding-box computation
