+++ dart/dynamics/Shape.cpp
@@ -42,6 +42,7 @@ namespace dynamics {
 //==============================================================================
 Shape::Shape(ShapeType type)
   : mBoundingBox(),
+    mIsBoundingBoxDirty(true),
     mVolume(0.0),
     mIsVolumeDirty(true),
     mID(mCounter++),
@@ -71,7 +72,10 @@ Shape::~Shape()
 //==============================================================================
 const math::BoundingBox& Shape::getBoundingBox() const
 {
-    return mBoundingBox;
+  if (mIsBoundingBoxDirty)
+    updateBoundingBox();
+
+  return mBoundingBox;
 }
 
 //==============================================================================
