+++ dart/dynamics/SphereShape.hpp
@@ -70,6 +70,9 @@ public:
 
 protected:
   // Documentation inherited.
+  void updateBoundingBox() const override;
+
+  // Documentation inherited.
   void updateVolume() const override;
 
 private:
