### 2018-01-19 1991f22
Put back functions in setMesh
