### 2018-01-19 #500 @betatim
I've not put in the time to get the software working/installed on my Mac. It is unfortunately also not anywhere near the top of my todo list. So if we can find someone else who wants to take over I would be A++ super Ok with that.
