### 2018-01-19 1fe0ce2
Fix oversights in lazy volume/bounding-box computation
