+++ dart/dynamics/CapsuleShape.cpp
@@ -133,6 +133,8 @@ void CapsuleShape::updateBoundingBox() const
 
   mBoundingBox.setMin(-corner);
   mBoundingBox.setMax(corner);
+
+  mIsBoundingBoxDirty = false;
 }
 
 //==============================================================================
