### 2018-01-19 536d720
Fix for #734 (#958)
