### 2018-01-19 fe2d49e
Fix another oversight in lazy volume/bounding-box computation
