### 2018-01-19 d22fde0
Lazy evaluation for shape volume and bounding-box (#959)
