+++ dart/dynamics/ConeShape.cpp
@@ -48,8 +48,6 @@ ConeShape::ConeShape(double radius, double height)
 {
   assert(0.0 < radius);
   assert(0.0 < height);
-  updateBoundingBoxDim();
-  updateVolume();
 }
 
 //==============================================================================
@@ -76,8 +74,8 @@ void ConeShape::setRadius(double radius)
 {
   assert(0.0 < radius);
   mRadius = radius;
-  updateBoundingBoxDim();
-  updateVolume();
+  mIsBoundingBoxDirty = true;
+  mIsVolumeDirty = true;
 }
 
 //==============================================================================
@@ -91,8 +89,8 @@ void ConeShape::setHeight(double height)
 {
   assert(0.0 < height);
   mHeight = height;
-  updateBoundingBoxDim();
-  updateVolume();
+  mIsBoundingBoxDirty = true;
+  mIsVolumeDirty = true;
 }
 
 //==============================================================================
@@ -117,24 +115,27 @@ Eigen::Matrix3d ConeShape::computeInertia(
 }
 
 //==============================================================================
-void ConeShape::updateVolume()
+void ConeShape::updateBoundingBox() const
+{
+  const Eigen::Vector3d corner(mRadius, mRadius, mRadius + 0.5*mHeight);
+
+  mBoundingBox.setMin(-corner);
+  mBoundingBox.setMax(corner);
+
+  mIsBoundingBoxDirty = false;
+}
+
+//==============================================================================
+void ConeShape::updateVolume() const
 {
   mVolume = computeVolume(mRadius, mHeight);
+  mIsVolumeDirty = false;
 }
 
 //==============================================================================
 Eigen::Matrix3d ConeShape::computeInertia(double mass) const
 {
   return computeInertia(mRadius, mHeight, mass);
-}
-
-//==============================================================================
-void ConeShape::updateBoundingBoxDim()
-{
-  const Eigen::Vector3d corner(mRadius, mRadius, mRadius + 0.5*mHeight);
-
-  mBoundingBox.setMin(-corner);
-  mBoundingBox.setMax(corner);
 }
 
 }  // namespace dynamics
