+++ dart/dynamics/SphereShape.hpp
@@ -70,7 +70,10 @@ public:
 
 protected:
   // Documentation inherited.
-  void updateVolume() override;
+  void updateBoundingBox() const override;
+
+  // Documentation inherited.
+  void updateVolume() const override;
 
 private:
   /// Radius of this Sphere
