+++ dart/dynamics/MeshShape.cpp
@@ -220,49 +220,58 @@ void MeshShape::setMesh(
   }
 
   mResourceRetriever = _resourceRetriever;
-
-  _updateBoundingBoxDim();
-  updateVolume();
 }
 
-void MeshShape::setScale(const Eigen::Vector3d& _scale) {
+//==============================================================================
+void MeshShape::setScale(const Eigen::Vector3d& _scale)
+{
   assert(_scale[0] > 0.0);
   assert(_scale[1] > 0.0);
   assert(_scale[2] > 0.0);
   mScale = _scale;
-  _updateBoundingBoxDim();
-  updateVolume();
+  mIsBoundingBoxDirty = true;
+  mIsVolumeDirty = true;
 }
 
-const Eigen::Vector3d& MeshShape::getScale() const {
+//==============================================================================
+const Eigen::Vector3d& MeshShape::getScale() const
+{
   return mScale;
 }
 
+//==============================================================================
 void MeshShape::setColorMode(ColorMode _mode)
 {
   mColorMode = _mode;
 }
 
+//==============================================================================
 MeshShape::ColorMode MeshShape::getColorMode() const
 {
   return mColorMode;
 }
 
+//==============================================================================
 void MeshShape::setColorIndex(int _index)
 {
   mColorIndex = _index;
 }
 
+//==============================================================================
 int MeshShape::getColorIndex() const
 {
   return mColorIndex;
 }
 
-int MeshShape::getDisplayList() const {
+//==============================================================================
+int MeshShape::getDisplayList() const
+{
   return mDisplayList;
 }
 
-void MeshShape::setDisplayList(int _index) {
+//==============================================================================
+void MeshShape::setDisplayList(int _index)
+{
   mDisplayList = _index;
 }
 
@@ -270,20 +279,17 @@ void MeshShape::setDisplayList(int _index) {
 Eigen::Matrix3d MeshShape::computeInertia(double _mass) const
 {
   // Use bounding box to represent the mesh
-  return BoxShape::computeInertia(mBoundingBox.computeFullExtents(), _mass);
+  return BoxShape::computeInertia(getBoundingBox().computeFullExtents(), _mass);
 }
 
-void MeshShape::updateVolume() {
-  Eigen::Vector3d bounds = mBoundingBox.computeFullExtents();
-  mVolume = bounds.x() * bounds.y() * bounds.z();
-}
-
-void MeshShape::_updateBoundingBoxDim() {
-
-  if(!mMesh)
+//==============================================================================
+void MeshShape::updateBoundingBox() const
+{
+  if (!mMesh)
   {
     mBoundingBox.setMin(Eigen::Vector3d::Zero());
     mBoundingBox.setMax(Eigen::Vector3d::Zero());
+    mIsBoundingBoxDirty = false;
     return;
   }
 
@@ -294,8 +300,10 @@ void MeshShape::_updateBoundingBoxDim() {
   double min_Y = std::numeric_limits<double>::infinity();
   double min_Z = std::numeric_limits<double>::infinity();
 
-  for (unsigned int i = 0; i < mMesh->mNumMeshes; i++) {
-    for (unsigned int j = 0; j < mMesh->mMeshes[i]->mNumVertices; j++) {
+  for (unsigned int i = 0; i < mMesh->mNumMeshes; i++)
+  {
+    for (unsigned int j = 0; j < mMesh->mMeshes[i]->mNumVertices; j++)
+    {
       if (mMesh->mMeshes[i]->mVertices[j].x > max_X)
         max_X = mMesh->mMeshes[i]->mVertices[j].x;
       if (mMesh->mMeshes[i]->mVertices[j].x < min_X)
[…]
