### 2018-01-22 9b790ab
[joss] Add more papers to paper.md
