+++ paper/paper.md
@@ -56,18 +56,21 @@ DART (Dynamic Animation and Robotics Toolkit) is a collaborative, cross-platform
 
 # Research publications utilizing DART
 
-1. Data-Driven Approach to Simulating Realistic Human Joint Constraints, Yifeng Jiang, and C. Karen Liu [[arXiv](https://arxiv.org/abs/1709.08685)]
-1. Multi-task Learning with Gradient Guided Policy Specialization, Wenhao Yu, Greg Turk, and C. Karen Liu [[arXiv](https://arxiv.org/abs/1709.07979)]
-1. Learning Human Behaviors for Robot-Assisted Dressing, Alex Clegg, Wenhao Yu, Jie Tan, Charlie Kemp, Greg Turk, and C. Karen Liu [[arXiv](https://arxiv.org/abs/1709.07033)]
-1. Expanding Motor Skills through Relay Neural Networks, Visak C.V. Kumar, Sehoon Ha, and C. Karen Liu [[arXiv](https://arxiv.org/abs/1709.07932)]
+1. Bayesian Optimization with Automatic Prior Selection for Data-Efficient Direct Policy Search, Rémi Pautrat, Konstantinos Chatzilygeroudis and Jean-Baptiste MouretRémi Pautrat, Konstantinos Chatzilygeroudis and Jean-Baptiste Mouret, in Proceedings of the International Conference on Robotics and Automation (ICRA), 2018 [[PDF](https://arxiv.org/pdf/1709.06919.pdf)]
+1. Data-Driven Approach to Simulating Realistic Human Joint Constraints, Yifeng Jiang, and C. Karen Liu, (preprint), 2017 [[arXiv](https://arxiv.org/abs/1709.08685)]
+1. Multi-task Learning with Gradient Guided Policy Specialization, Wenhao Yu, Greg Turk, and C. Karen Liu, (preprint), 2017 [[arXiv](https://arxiv.org/abs/1709.07979)]
+1. Learning Human Behaviors for Robot-Assisted Dressing, Alex Clegg, Wenhao Yu, Jie Tan, Charlie Kemp, Greg Turk, and C. Karen Liu, (preprint), 2017 [[arXiv](https://arxiv.org/abs/1709.07033)]
+1. Expanding Motor Skills through Relay Neural Networks, Visak C.V. Kumar, Sehoon Ha, and C. Karen Liu, (preprint), 2017 [[arXiv](https://arxiv.org/abs/1709.07932)]
+1. Reset-free Trial-and-Error Learning for Robot Damage Recovery, Konstantinos Chatzilygeroudis, Vassilis Vassiliades, Jean-Baptiste Mouret, Robotics and Autonomous Systems, 2017 [[arXiv](https://arxiv.org/pdf/1610.04213.pdf)]
 1. Stair Negotiation Made Easier Using Novel Interactive Energy-Recycling Assistive Stairs, Yun Seong Song, Sehoon Ha, Hsiang Hsu, Lena H. Ting, and C. Karen Liu, in PLOS ONE, 2017 [[PDF](http://journals.plos.org/plosone/article?id=10.1371/journal.pone.0179637)]
 1. Learning a Unified Control Policy for Safe Falling, Visak C.V. Kumar, Sehoon Ha, and C. Karen Liu, in IEEE/RSJ International Conference on Intelligent Robots and Systems (IROS), 2017 [[PDF](https://arxiv.org/abs/1703.02905)]
[…]
