### 2018-01-22 #500 @jslee02
Three more papers were added to the "Research publications utilizing DART" section by https://github.com/dartsim/dart/issues/957.
