### 2018-01-23 a416e28
Use homebrew/libccd instead of dartsim/dart/libccd
https://github.com/Homebrew/homebrew-core/pull/22359
