+++ ci/install_osx.sh
@@ -11,7 +11,7 @@ brew install boost | grep -v '%$'
 brew install eigen | grep -v '%$'
 brew install tinyxml | grep -v '%$'
 brew install tinyxml2 | grep -v '%$'
-brew install dartsim/dart/libccd | grep -v '%$'
+brew install libccd | grep -v '%$'
 brew install nlopt | grep -v '%$'
 brew install dartsim/dart/ipopt | grep -v '%$'
 brew install ros/deps/urdfdom | grep -v '%$'
