### 2018-01-24 dartsim/dart#961 @costashatz
FCL is too slow?
Hello,

#### Environment
* Select DART version: master and 6.3.0
* Select OS name and version name: Ubuntu 16.04
* Select compiler name and version number: GCC 5.4.0
* FCL version: tried both 0.3.2 and 0.5.0

#### Current Behavior
I am using the dominoes tutorial and I create several of them (like 8-12), I push the first domino and the simulation is really slow when the collisions start to happen (i.e., really really bad)! I altered a bit the tutorial to not include the arm (which has meshes for collision) and switch to the simple `DARTCollisionDetector` and now I can put much much more dominoes (throw them from random heights/positions with collisions etc) and the simulation is really smooth. I tried the default FCL in my machine (0.3.2) and the latest (stable) version (0.5.0), but nothing changed.

Is this expected? Or am I missing something (for example, ways of making the collision checking faster!)?

Thanks!
