### 2018-01-25 44faf02
Update .appveyor.yml
