+++ .appveyor.yml
@@ -67,7 +67,6 @@ install:
 # preserve contents of selected directories and files across project builds
 cache:
   - C:\tools\vcpkg\installed -> .appveyor.yml
-  - C:\tools\vcpkg\packages -> .appveyor.yml
 
 build:
   project: C:\projects\dart\build\dart.sln # path to Visual Studio solution or project
