+++ .appveyor.yml
@@ -45,16 +45,16 @@ install:
   - vcpkg install fcl:%VCPKG_ARCH%
   - vcpkg install assimp:%VCPKG_ARCH%
   # ---- optional dependencies: nlopt, ipopt, bullet3, ode, flann, tinyxml2, urdfdom_headers, urdfdom, opengl, freeglut, osg
-  #- vcpkg install nlopt:%VCPKG_ARCH%
-  #- vcpkg install ipopt:%VCPKG_ARCH%
-  #- vcpkg install bullet3:%VCPKG_ARCH%
-  #- vcpkg install ode:%VCPKG_ARCH%
-  #- vcpkg install flann:%VCPKG_ARCH%
-  #- vcpkg install tinyxml2:%VCPKG_ARCH%
-  #- vcpkg install urdfdom-headers:%VCPKG_ARCH%
-  #- vcpkg install urdfdom:%VCPKG_ARCH%
-  #- vcpkg install opengl:%VCPKG_ARCH%
-  #- vcpkg install freeglut:%VCPKG_ARCH%
+  - vcpkg install nlopt:%VCPKG_ARCH%
+  - vcpkg install ipopt:%VCPKG_ARCH%
+  - vcpkg install bullet3:%VCPKG_ARCH%
+  - vcpkg install ode:%VCPKG_ARCH%
+  - vcpkg install flann:%VCPKG_ARCH%
+  - vcpkg install tinyxml2:%VCPKG_ARCH%
+  - vcpkg install urdfdom-headers:%VCPKG_ARCH%
+  - vcpkg install urdfdom:%VCPKG_ARCH%
+  - vcpkg install opengl:%VCPKG_ARCH%
+  - vcpkg install freeglut:%VCPKG_ARCH%
   #- vcpkg install osg:%VCPKG_ARCH%
   
   - vcpkg integrate install
