### 2018-01-25 a846748
Update .appveyor.yml
