+++ .appveyor.yml
@@ -74,4 +74,4 @@ build:
   verbosity: quiet                         # MSBuild verbosity level (quiet|minimal|normal|detailed)
 
 #test_script:
-  #- cmd: ctest --build-config %configuration% --parallel 4 --output-on-failure
+  - cmd: ctest --build-config %configuration% --parallel 4 --output-on-failure
