### 2018-01-25 50a4975
Update .appveyor.yml
