### 2018-01-25 #500 @jslee02
@Kevin-Mattheus-Moerman It seems @betatim is not available to review this submission at this moment (https://github.com/openjournals/joss-reviews/issues/500#issuecomment-358904542). Please let me know if there is anything I can do for proceeding.
