### 2018-01-25 #500 @costashatz
@Kevin-Mattheus-Moerman I could volunteer to review the library (if a second reviewer is needed). I've been using quite a lot, but I have no connection with the developers (I am not sure about the conflict of interest policy). Cheers!
