### 2018-01-25 #500 @Kevin-Mattheus-Moerman
@costashatz yes that would be great! I'll add a form for you to tick boxes.
