### 2018-01-25 #500 @costashatz
> @costashatz thanks again for jumping in to save the day  ! Could you give me an indication as to when you expect to finalize the review? Thanks

No problem! Since I am already using the library (i.e., I have it installed, already done the tutorials etc), I think it'll be quick!
