### 2018-01-25 #500 @costashatz
@jslee02  Great library and submission! Only two small remarks:

- Is there a specific reason why you do not have API docs for v6.3.0 that is being submitted? I am aware the API didn't change much from v6.1.1 to v6.3.0, but I'd guess it's not difficult to generate it right?
- Regarding the performance, as I already noted in [this issue](https://github.com/dartsim/dart/issues/961) there seems to be something slowing down the simulation with collisions (mainly the one of FCL) even with simple cases (like the collision/dominoes tutorial). Have you identified what the problem is? Does it come from DART or FCL?

Both of my remarks are minor as the library is very well written and documented!

Thanks!
