### 2018-01-26 4cf4e20
Update .appveyor.yml
