### 2018-01-26 b746849
Update .appveyor.yml
