+++ .appveyor.yml
@@ -50,9 +50,9 @@ install:
   #- vcpkg install bullet3:%VCPKG_ARCH%
   #- vcpkg install ode:%VCPKG_ARCH%
   #- vcpkg install flann:%VCPKG_ARCH%
-  - vcpkg install tinyxml2:%VCPKG_ARCH%
-  - vcpkg install urdfdom-headers:%VCPKG_ARCH%
-  - vcpkg install urdfdom:%VCPKG_ARCH%
+  #- vcpkg install tinyxml2:%VCPKG_ARCH%
+  #- vcpkg install urdfdom-headers:%VCPKG_ARCH%
+  #- vcpkg install urdfdom:%VCPKG_ARCH%
   #- vcpkg install opengl:%VCPKG_ARCH%
   #- vcpkg install freeglut:%VCPKG_ARCH%
   #- vcpkg install osg:%VCPKG_ARCH%
