### 2018-01-26 #500 @jslee02
@costashatz Thank you for the quick feedback. Please find below my answers to your remarks:

* Updated API docs can be found on the [website](http://dartsim.github.io/api.html) or [API repo](https://github.com/dartsim/api).

* Regarding the collision/contact performance, I [replied](https://github.com/dartsim/dart/issues/961#issuecomment-360874612) on the issue:

> This is somewhat an expected result. There is two main reasons behind this: (1) collision handling method in DART and (2) number of contact points generated by the collision detector.
>
> (1) DART uses rather slow but more precise contact constraint solver.
>
> AFAIK, there are two major categories of solving non-penetration contact constraint handling methods: (a) [LCP](https://en.wikipedia.org/wiki/Linear_complementarity_problem)-based method and (b) iterative method. I believe DART initially chose (a) method to compute precise contact points sacrificing speed, but we are working on supporting (b) for interactive simulations like manipulation in VR. [This](https://github.com/dartsim/dart/tree/enhancement/constraint_solver) is my working branch hoping to finish initial implementation by next couple of months.
>
> (2) `FCLCollisionDetector` generally returns more contact points than `DARTCollisionDetector`.
>
> In FCL setting of DART, we use meshes even for primitive shapes (e.g., box and sphere) by default, and this generally results in more contact points than primitive-primitive shape pair contacts (like `DARTCollisionDetector`). This is because FCL detects all the contact points of the meshes' triangle pairs in contact while primitive-primitive collision algorithm returns minimally required contacts (this can be varied by the concrete primitive collision algorithm). For example, box-box collision algorithm returns up to four contacts.
>
> Loosely speaking, the complexity of (a) method is O(n^3) where n is the number of contacts (assuming no friction). That said, more contacts make the simulate slow quickly.
>
> The reason we don't use FCL's primitive shape support is that it has a few unresolved problems. I had worked on addressing those issues (e.g., https://github.com/flexible-collision-library/fcl/pull/52, https://github.com/flexible-collision-library/fcl/pull/61), but there are some remains.
>
> There is persistent contact manifold technic of Bullet Physics engine that can address this issue. It manages the contact points reported from the collision detector to be less than 4, which is minimal contact points to support the contact shapes each other. I'm also considering having the similar feature in DART.
