### 2018-01-27 be36bcc
Add World::create()
