+++ dart/simulation/World.hpp
@@ -86,6 +86,9 @@ public:
   // Constructor and Destructor
   //--------------------------------------------------------------------------
 
+  /// Creates a World
+  static std::shared_ptr<World> create(const std::string& name);
+
   /// Constructor
   World(const std::string& _name = "world");
 
