### 2018-01-27 7c7ba76
Add default name parameter for World::create()
