+++ dart/simulation/World.hpp
@@ -87,7 +87,7 @@ public:
   //--------------------------------------------------------------------------
 
   /// Creates a World
-  static std::shared_ptr<World> create(const std::string& name);
+  static std::shared_ptr<World> create(const std::string& name = "world");
 
   /// Constructor
   World(const std::string& _name = "world");
