### 2018-01-27 2d5d016
Use World::create() in the existing codebase
