### 2018-01-28 c65d787
Add World::create() (#962)
