### 2018-01-31 #500 @costashatz
> @costashatz Thank you for the quick feedback. Please find below my answers to your remarks:
>
> Updated API docs can be found on the website or API repo.
>
> Regarding the collision/contact performance, I replied on the issue

@jslee02  Thank you for your fast answers (I will continue the discussion in the issue regarding the collision/contact performance). I am very much satisfied by the library and the submission! Best of luck!

@Kevin-Mattheus-Moerman Let me know if you need me to do anything else..
