### 2018-02-01 6ef54f9
Update changelog item orders (descending in the same category)
