### 2018-02-01 e65f57e
Update changelog for #964 (#966)
