### 2018-02-01 fec4c24
Port clang warning fix over to 6.4 (#964)
