+++ README.md
@@ -5,6 +5,6 @@ Visit the [DART website](http://dartsim.github.io/) for more information
 * [Gallery](http://dartsim.github.io/gallery.html)
 * Installation: [Ubuntu](http://dartsim.github.io/install_dart_on_ubuntu.html) | [macOS](http://dartsim.github.io/install_dart_on_mac.html) | [Windows](http://dartsim.github.io/install_dart_on_windows.html)
 * [Tutorials](http://dartsim.github.io/tutorials_introduction.html)
-* [API Documentation](http://dartsim.github.io/dart/)
+* [API Documentation](http://dartsim.github.io/dart/v6.3.0/)
 * Python bindings: [PyDart2](https://github.com/sehoonha/pydart2), [dartpy](https://github.com/personalrobotics/dartpy) (experimental)
 * OpenAI Gym with DART support: [DartEnv](https://github.com/DartEnv/dart-env)
