### 2018-02-04 627d1f9
Add Codacy badge to README.md
