+++ README.md
@@ -1,5 +1,5 @@
 # DART: Dynamic Animation and Robotics Toolkit
-[![Build Status](https://travis-ci.org/dartsim/dart.png?branch=master)](https://travis-ci.org/dartsim/dart) [![Build status](https://ci.appveyor.com/api/projects/status/6rta8olo95bpu84r/branch/master?svg=true)](https://ci.appveyor.com/project/jslee02/dart/branch/master) [![codecov](https://codecov.io/gh/dartsim/dart/branch/master/graph/badge.svg)](https://codecov.io/gh/dartsim/dart)
+[![Build Status](https://travis-ci.org/dartsim/dart.png?branch=master)](https://travis-ci.org/dartsim/dart) [![Build status](https://ci.appveyor.com/api/projects/status/6rta8olo95bpu84r/branch/master?svg=true)](https://ci.appveyor.com/project/jslee02/dart/branch/master) [![codecov](https://codecov.io/gh/dartsim/dart/branch/master/graph/badge.svg)](https://codecov.io/gh/dartsim/dart) [![Codacy Badge](https://api.codacy.com/project/badge/Grade/1251beff30164eb7ba60a1cc9e6daef9)](https://www.codacy.com/app/jslee02/dart?utm_source=github.com&amp;utm_medium=referral&amp;utm_content=dartsim/dart&amp;utm_campaign=Badge_Grade)
 
 Visit the [DART website](http://dartsim.github.io/) for more information
 * [Gallery](http://dartsim.github.io/gallery.html)
