### 2018-02-04 b2270f9
Implement getFilePath for DartResourceRetriever
