### 2018-02-05 deaa6f3
Revert change of PackageResourceRetriever::exists()
