### 2018-02-05 9e93be9
Add osgValkyrie example
