+++ examples/osgExamples/osgValkyrie/Configuration.h
@@ -1,0 +1,6 @@
+#ifndef PATH_H
+#define PATH_H
+
+#define THIS_COM "/home/jslee02/dev/prl/dart/dartDebug/"
+
+#endif
