### 2018-02-05 450827c
Render texture images when available
