+++ dart/dynamics/MeshShape.cpp
@@ -127,24 +127,24 @@ aiMaterial::~aiMaterial()
 namespace dart {
 namespace dynamics {
 
-MeshShape::MeshShape(const Eigen::Vector3d& _scale, const aiScene* _mesh,
-                     const std::string& _path,
-                     const common::ResourceRetrieverPtr& _resourceRetriever)
+//==============================================================================
+MeshShape::MeshShape(
+    const Eigen::Vector3d& scale,
+    const aiScene* mesh,
+    const common::Uri& path,
+    common::ResourceRetrieverPtr resourceRetriever)
   : Shape(MESH),
-    mResourceRetriever(_resourceRetriever),
     mDisplayList(0),
     mColorMode(MATERIAL_COLOR),
     mColorIndex(0)
 {
-  assert(_scale[0] > 0.0);
-  assert(_scale[1] > 0.0);
-  assert(_scale[2] > 0.0);
-
-  setMesh(_mesh, _path, _resourceRetriever);
-  setScale(_scale);
+  setMesh(mesh, path, std::move(resourceRetriever));
+  setScale(scale);
 }
 
-MeshShape::~MeshShape() {
+//==============================================================================
+MeshShape::~MeshShape()
+{
   delete mMesh;
 }
 
@@ -161,15 +161,25 @@ const std::string& MeshShape::getStaticType()
   return type;
 }
 
-const aiScene* MeshShape::getMesh() const {
+//==============================================================================
+const aiScene* MeshShape::getMesh() const
+{
   return mMesh;
 }
 
-const std::string& MeshShape::getMeshUri() const
+//==============================================================================
+std::string MeshShape::getMeshUri() const
+{
+  return mMeshUri.toString();
+}
+
+//==============================================================================
+const common::Uri& MeshShape::getMeshUri2() const
 {
   return mMeshUri;
 }
 
+//==============================================================================
 void MeshShape::update()
 {
   // Do nothing
@@ -186,49 +196,58 @@ void MeshShape::notifyAlphaUpdated(double alpha)
   }
 }
 
+//==============================================================================
 const std::string& MeshShape::getMeshPath() const
 {
   return mMeshPath;
 }
 
-void MeshShape::setMesh(
-  const aiScene* _mesh, const std::string& _path,
-  const common::ResourceRetrieverPtr& _resourceRetriever)
+//==============================================================================
+common::ResourceRetrieverPtr MeshShape::getResourceRetriever()
 {
-  mMesh = _mesh;
+  return mResourceRetriever;
+}
 
-  if(nullptr == _mesh) {
-    mMeshPath = "";
-    mMeshUri = "";
+//==============================================================================
+void MeshShape::setMesh(
+  const aiScene* mesh, const std::string& path,
+  common::ResourceRetrieverPtr resourceRetriever)
+{
+  setMesh(mesh, common::Uri(path), std::move(resourceRetriever));
+}
+
+//==============================================================================
+void MeshShape::setMesh(
+  const aiScene* mesh,
+  const common::Uri& uri,
+  common::ResourceRetrieverPtr resourceRetriever)
+{
+  mMesh = mesh;
+
+  if (!mMesh)
+  {
+    mMeshUri.clear();
+    mMeshPath.clear();
     mResourceRetriever = nullptr;
     return;
   }
 
-  common::Uri uri;
-  if(uri.fromString(_path))
-  {
-    mMeshUri = _path;
+  mMeshUri = uri;
 
-    if(uri.mScheme.get_value_or("file") == "file")
-      mMeshPath = uri.mPath.get_value_or("");
-  }
+  if (resourceRetriever)
+    mMeshPath = resourceRetriever->getFilePath(uri);
   else
-  {
[…]
