### 2018-02-05 6dd3203
Add osgValkyrie example (reverted from commit 9e93be971557e7a93c0917050aea4779ff0b3aca)
