### 2018-02-05 2e0d73d
Fix tests of invalid URIs
