+++ unittests/unit/test_PackageResourceRetriever.cpp
@@ -31,6 +31,7 @@
  */
 
 #include <gtest/gtest.h>
+#include "dart/config.hpp"
 #include "dart/utils/PackageResourceRetriever.hpp"
 #include "TestHelpers.hpp"
 
@@ -55,14 +56,14 @@ TEST(PackageResourceRetriever, exists_DelegateFails_ReturnsFalse)
 {
   // GTest breaks the string concatenation.
 #ifdef _WIN32
-  const char* expected = "file:///" "dart://sample/test/foo";
+  const char* expected = "file:///" DART_DATA_LOCAL_PATH"test/foo";
 #else
-  const char* expected = "file://" "dart://sample/test/foo";
+  const char* expected = "file://" DART_DATA_LOCAL_PATH"test/foo";
 #endif
 
   auto mockRetriever = std::make_shared<AbsentResourceRetriever>();
   PackageResourceRetriever retriever(mockRetriever);
-  retriever.addPackageDirectory("test", "dart://sample/test");
+  retriever.addPackageDirectory("test", DART_DATA_LOCAL_PATH"test");
 
   EXPECT_FALSE(retriever.exists(Uri::createFromString("package://test/foo")));
   ASSERT_EQ(1u, mockRetriever->mExists.size());
@@ -75,7 +76,7 @@ TEST(PackageResourceRetriever, exists_UnsupportedUri_ReturnsFalse)
 {
   auto mockRetriever = std::make_shared<PresentResourceRetriever>();
   PackageResourceRetriever retriever(mockRetriever);
-  retriever.addPackageDirectory("test", "dart://sample/test");
+  retriever.addPackageDirectory("test", DART_DATA_LOCAL_PATH"test");
 
   EXPECT_FALSE(retriever.exists(Uri::createFromString("foo://test/foo")));
   EXPECT_TRUE(mockRetriever->mExists.empty());
@@ -86,14 +87,14 @@ TEST(PackageResourceRetriever, exists_UnsupportedUri_ReturnsFalse)
 TEST(PackageResourceRetriever, exists_StripsTrailingSlash)
 {
 #ifdef _WIN32
-  const char* expected = "file:///" "dart://sample/test/foo";
+  const char* expected = "file:///" DART_DATA_LOCAL_PATH"test/foo";
 #else
-  const char* expected = "file://" "dart://sample/test/foo";
+  const char* expected = "file://" DART_DATA_LOCAL_PATH"test/foo";
 #endif
 
   auto mockRetriever = std::make_shared<PresentResourceRetriever>();
   PackageResourceRetriever retriever(mockRetriever);
-  retriever.addPackageDirectory("test", "dart://sample/test/");
+  retriever.addPackageDirectory("test", DART_DATA_LOCAL_PATH"test/");
 
   EXPECT_TRUE(retriever.exists(Uri::createFromString("package://test/foo")));
   ASSERT_EQ(1u, mockRetriever->mExists.size());
@@ -105,15 +106,15 @@ TEST(PackageResourceRetriever, exists_StripsTrailingSlash)
 TEST(PackageResourceRetriever, exists_FirstUriSucceeds)
 {
 #ifdef _WIN32
-  const char* expected = "file:///" "dart://sample/test1/foo";
+  const char* expected = "file:///" DART_DATA_LOCAL_PATH"test1/foo";
 #else
-  const char* expected = "file://" "dart://sample/test1/foo";
+  const char* expected = "file://" DART_DATA_LOCAL_PATH"test1/foo";
 #endif
 
   auto mockRetriever = std::make_shared<PresentResourceRetriever>();
   PackageResourceRetriever retriever(mockRetriever);
[…]
