### 2018-02-05 c79de2d
Add comment on why an additional slash is required for Windows
