### 2018-02-05 0127265
Use Brewfile to install depedencies on macOS
