### 2018-02-05 42349b9
Use tapping instead of using full path
It seems brew bundle doesn't like using full path: https://travis-ci.org/dartsim/dart/jobs/337595721#L65-L71
