+++ ci/install_osx.sh
@@ -1,2 +1,4 @@
 brew update > /dev/null
+brew tap dartsim/dart
+brew tap ros/deps
 brew bundle --file="$TRAVIS_BUILD_DIR/ci/Brewfile"
